requirements.txt
django
djangorestframework
pycrypto
django-cronjobs
